extract out the dismiss thing
